Messages storage data structure fix. UI fixes.

+++ WebSockP2P/WebSockP2P/Screens/DevicesListTableCell.m
@@ -28,7 +28,7 @@
 }
 
 - (void)setMessagesCount:(NSUInteger)count {
-    self.messagesCountLabel.text = [NSString stringWithFormat:@"%lu", count];
+    self.messagesCountLabel.text = [NSString stringWithFormat:@"%lu", (unsigned long)count];
 }
 
 @end
